requirements.txt
aniso8601==10.0.1
blinker==1.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
Flask==3.1.2
Flask-Logging==0.1.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
oracledb==3.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
win32_setctime==1.2.0
